requirements.txt
requests>=2.32
openai>=2.18
pymupdf4llm>=0.2.9
PyMuPDF>=1.26
python-pptx>=1.0
Jinja2>=3.1
evaluate>=0.4
rouge-score
torch>=2.9
transformers>=4.57
torchvision>=0.24
scipy>=1.15
Pillow>=12.0
